Fix movement mechanics, move stuff around
